+++ Sources/PDFjet/Table.swift
@@ -487,7 +487,9 @@ public class Table : Drawable {
                     j += 1
                 }
                 if page != nil {
-                    page!.setBrushColor(cell.getTextColor())
+                    if cell.textColor != Cell.NO_COLOR {
+                        page!.setBrushColor(cell.textColor)
+                    }
                     if i == (numOfHeaderRows - 1) {
                         cell.setBorder(Border.BOTTOM, true)
                     }
@@ -524,7 +526,9 @@ public class Table : Drawable {
                     i += 1
                 }
                 if page != nil {
-                    page!.setBrushColor(cell.getTextColor())
+                    if cell.textColor != Cell.NO_COLOR {
+                        page!.setBrushColor(cell.textColor)
+                    }
                     cell.drawOn(page!, x, y, w, h)
                 }
                 x += w
@@ -768,10 +772,10 @@ public class Table : Drawable {
                     cell2.setWidth(cell.getWidth())
                     cell2.setLeftPadding(cell.getLeftPadding())
                     cell2.setRightPadding(cell.getRightPadding())
-                    cell2.setBackgroundColor(cell.getBackgroundColor())
+                    cell2.backgroundColor = cell.backgroundColor
                     cell2.setBorderWidth(cell.getBorderWidth())
-                    cell2.setBorderColor(cell.getBorderColor())
-                    cell2.setTextColor(cell.getTextColor())
+                    cell2.borderColor = cell.borderColor
+                    cell2.textColor = cell.textColor
                     cell2.setColSpan(cell.getColSpan())
                     cell2.setTextAlignment(cell.getTextAlignment())
                     cell2.setBorder(Border.TOP, cell.getBorder(Border.TOP))